feat: implements move-on direction class
The main part of the rover's movement relies on executing commands
for the directions, and given the directions, it needs to move
along the x and y axes.

I decided instead to use a command structure, or as an anonymous
function, not to make it complex, and to make testing easier and in
turn reduce coupling.